refactor: move registerChain helper to chain-info
out of exo module
