+++ packages/orchestration/src/chain-info.js
@@ -1,7 +1,8 @@
-import { registerChain } from './exos/chain-hub.js';
-
-// Refresh with scripts/refresh-chain-info.ts
-import fetchedChainInfo from './fetched-chain-info.js';
+import { E } from '@endo/far';
+import { mustMatch } from '@endo/patterns';
+import { connectionKey } from './exos/chain-hub.js';
+import fetchedChainInfo from './fetched-chain-info.js'; // Refresh with scripts/refresh-chain-info.ts
+import { CosmosChainInfoShape } from './typeGuards.js';
 
 /** @import {CosmosChainInfo, EthChainInfo} from './types.js'; */
 
@@ -63,6 +64,44 @@ const knownChains = /** @satisfies {Record<string, ChainInfo>} */ (
 );
 
 /** @typedef {typeof knownChains} KnownChains */
+
+/**
+ * @param {ERef<import('@agoric/vats').NameHubKit['nameAdmin']>} agoricNamesAdmin
+ * @param {string} name
+ * @param {CosmosChainInfo} chainInfo
+ * @param {(...messages: string[]) => void} log
+ */
+export const registerChain = async (
+  agoricNamesAdmin,
+  name,
+  chainInfo,
+  log = () => {},
+) => {
+  const { nameAdmin } = await E(agoricNamesAdmin).provideChild('chain');
+  const { nameAdmin: connAdmin } =
+    await E(agoricNamesAdmin).provideChild('chainConnection');
+
+  mustMatch(chainInfo, CosmosChainInfoShape);
+  const { connections = {}, ...vertex } = chainInfo;
+
+  const promises = [
+    E(nameAdmin)
+      .update(name, vertex)
+      .then(() => log(`registered agoricNames chain.${name}`)),
+  ];
+
+  // FIXME updates redundantly, twice per edge
+  for (const [counterChainId, connInfo] of Object.entries(connections)) {
+    const key = connectionKey(chainInfo.chainId, counterChainId);
+    promises.push(
+      E(connAdmin)
+        .update(key, connInfo)
+        .then(() => log(`registering agoricNames chainConnection.${key}`)),
+    );
+  }
+  // Bundle to pipeline IO
+  await Promise.all(promises);
+};
 
 /**
  * @param {ERef<import('@agoric/vats').NameHubKit['nameAdmin']>} agoricNamesAdmin
